+++ internal/engine/engine.go
@@ -854,107 +854,6 @@ func (e *Engine) executeSingleLineShell(shellStmt *ast.ShellStatement, ctx *Exec
 	return nil
 }
 
-// executeTry executes a try/catch/finally statement
-// executeDocker executes Docker operations
-func (e *Engine) executeDocker(dockerStmt *ast.DockerStatement, ctx *ExecutionContext) error {
-	// Interpolate variables in Docker statement
-	operation := dockerStmt.Operation
-	resource := dockerStmt.Resource
-	name := e.interpolateVariables(dockerStmt.Name, ctx)
-
-	// Interpolate options
-	options := make(map[string]string, len(dockerStmt.Options))
-	for key, value := range dockerStmt.Options {
-		options[key] = e.interpolateVariables(value, ctx)
-	}
-
-	if e.dryRun {
-		return e.buildDockerCommand(operation, resource, name, options, true)
-	}
-
-	// Show what we're about to do with appropriate emoji
-	switch operation {
-	case "build":
-		_, _ = fmt.Fprintf(e.output, "🔨 Building Docker image")
-		if name != "" {
-			_, _ = fmt.Fprintf(e.output, ": %s", name)
-		}
-		_, _ = fmt.Fprintf(e.output, "\n")
-	case "push":
-		_, _ = fmt.Fprintf(e.output, "📤 Pushing Docker image")
-		if name != "" {
-			_, _ = fmt.Fprintf(e.output, ": %s", name)
-		}
-		if registry, exists := options["to"]; exists {
-			_, _ = fmt.Fprintf(e.output, " to %s", registry)
-		}
-		_, _ = fmt.Fprintf(e.output, "\n")
-	case "pull":
-		_, _ = fmt.Fprintf(e.output, "📥 Pulling Docker image")
-		if name != "" {
-			_, _ = fmt.Fprintf(e.output, ": %s", name)
-		}
-		_, _ = fmt.Fprintf(e.output, "\n")
-	case "run":
-		_, _ = fmt.Fprintf(e.output, "🚀 Running Docker container")
-		if name != "" {
-			_, _ = fmt.Fprintf(e.output, ": %s", name)
-		}
-		if port, exists := options["port"]; exists {
-			_, _ = fmt.Fprintf(e.output, " on port %s", port)
-		}
-		_, _ = fmt.Fprintf(e.output, "\n")
-	case "stop":
-		_, _ = fmt.Fprintf(e.output, "🛑 Stopping Docker container")
-		if name != "" {
-			_, _ = fmt.Fprintf(e.output, ": %s", name)
-		}
-		_, _ = fmt.Fprintf(e.output, "\n")
-	case "remove":
-		_, _ = fmt.Fprintf(e.output, "🗑️  Removing Docker %s", resource)
-		if name != "" {
-			_, _ = fmt.Fprintf(e.output, ": %s", name)
-		}
-		_, _ = fmt.Fprintf(e.output, "\n")
-	case "compose":
-		command := options["command"]
-		switch command {
-		case "up":
-			_, _ = fmt.Fprintf(e.output, "🚀 Starting Docker Compose services\n")
-		case "down":
-			_, _ = fmt.Fprintf(e.output, "🛑 Stopping Docker Compose services\n")
-		case "build":
-			_, _ = fmt.Fprintf(e.output, "🔨 Building Docker Compose services\n")
-		default:
-			_, _ = fmt.Fprintf(e.output, "🐳 Running Docker Compose: %s\n", command)
-		}
-	case "scale":
-		if resource == "compose" {
-			replicas := options["replicas"]
-			_, _ = fmt.Fprintf(e.output, "📊 Scaling Docker Compose service")
-			if name != "" {
-				_, _ = fmt.Fprintf(e.output, " %s", name)
-			}
-			if replicas != "" {
-				_, _ = fmt.Fprintf(e.output, " to %s replicas", replicas)
-			}
-			_, _ = fmt.Fprintf(e.output, "\n")
-		}
-	default:
-		_, _ = fmt.Fprintf(e.output, "🐳 Running Docker %s", operation)
-		if resource != "" {
-			_, _ = fmt.Fprintf(e.output, " %s", resource)
-		}
-		if name != "" {
-			_, _ = fmt.Fprintf(e.output, " %s", name)
-		}
-		_, _ = fmt.Fprintf(e.output, "\n")
-	}
-
-	// Build and execute the actual command
-	return e.buildDockerCommand(operation, resource, name, options, false)
-}
-
 // buildDockerCommand builds and optionally executes the Docker command
 func (e *Engine) buildDockerCommand(operation, resource, name string, options map[string]string, dryRun bool) error {
 	var dockerCmd []string
@@ -1021,110 +920,6 @@ func (e *Engine) buildDockerCommand(operation, resource, name string, options ma
 	// return cmd.Run()
 
 	return nil
-}
-
-// executeGit executes Git operations
-func (e *Engine) executeGit(gitStmt *ast.GitStatement, ctx *ExecutionContext) error {
-	// Interpolate variables in Git statement
-	operation := gitStmt.Operation
-	resource := gitStmt.Resource
-	name := e.interpolateVariables(gitStmt.Name, ctx)
-
-	// Interpolate options
-	options := make(map[string]string, len(gitStmt.Options))
-	for key, value := range gitStmt.Options {
-		options[key] = e.interpolateVariables(value, ctx)
-	}
-
-	if e.dryRun {
-		return e.buildGitCommand(operation, resource, name, options, true)
-	}
-
-	// Show what we're about to do with appropriate emoji
-	switch operation {
-	case "create":
-		switch resource {
-		case "branch":
-			_, _ = fmt.Fprintf(e.output, "🌿 Creating Git branch")
-			if name != "" {
-				_, _ = fmt.Fprintf(e.output, ": %s", name)
-			}
-		case "tag":
-			_, _ = fmt.Fprintf(e.output, "🏷️  Creating Git tag")
-			if name != "" {
-				_, _ = fmt.Fprintf(e.output, ": %s", name)
-			}
-		}
-		_, _ = fmt.Fprintf(e.output, "\n")
-	case "checkout":
-		_, _ = fmt.Fprintf(e.output, "🔀 Checking out Git branch")
-		if name != "" {
-			_, _ = fmt.Fprintf(e.output, ": %s", name)
-		}
-		_, _ = fmt.Fprintf(e.output, "\n")
-	case "merge":
-		_, _ = fmt.Fprintf(e.output, "🔀 Merging Git branch")
-		if name != "" {
-			_, _ = fmt.Fprintf(e.output, ": %s", name)
-		}
-		_, _ = fmt.Fprintf(e.output, "\n")
-	case "commit":
-		_, _ = fmt.Fprintf(e.output, "💾 Committing Git changes")
-		if message, exists := options["message"]; exists {
-			_, _ = fmt.Fprintf(e.output, ": %s", message)
-		}
-		_, _ = fmt.Fprintf(e.output, "\n")
-	case "push":
-		if resource == "tag" {
-			_, _ = fmt.Fprintf(e.output, "📤 Pushing Git tag")
-			if name != "" {
-				_, _ = fmt.Fprintf(e.output, ": %s", name)
-			}
-		} else {
-			_, _ = fmt.Fprintf(e.output, "📤 Pushing Git changes")
-			if remote, exists := options["remote"]; exists {
-				_, _ = fmt.Fprintf(e.output, " to %s", remote)
-			}
-			if branch, exists := options["branch"]; exists {
-				_, _ = fmt.Fprintf(e.output, "/%s", branch)
-			}
-		}
-		_, _ = fmt.Fprintf(e.output, "\n")
-	case "clone":
-		_, _ = fmt.Fprintf(e.output, "📥 Cloning Git repository")
-		if name != "" {
-			_, _ = fmt.Fprintf(e.output, ": %s", name)
-		}
-		_, _ = fmt.Fprintf(e.output, "\n")
-	case "init":
-		_, _ = fmt.Fprintf(e.output, "🆕 Initializing Git repository\n")
-	case "add":
-		_, _ = fmt.Fprintf(e.output, "➕ Adding files to Git")
-		if name != "" {
-			_, _ = fmt.Fprintf(e.output, ": %s", name)
-		}
-		_, _ = fmt.Fprintf(e.output, "\n")
-	case "status":
-		_, _ = fmt.Fprintf(e.output, "📊 Checking Git status\n")
-	case "show":
-		if resource == "branch" {
-			_, _ = fmt.Fprintf(e.output, "🌿 Showing current Git branch\n")
-		} else {
-			_, _ = fmt.Fprintf(e.output, "📖 Showing Git information\n")
-		}
-	default:
-		_, _ = fmt.Fprintf(e.output, "🔗 Running Git %s", operation)
-		if resource != "" {
-			_, _ = fmt.Fprintf(e.output, " %s", resource)
-		}
-		if name != "" {
-			_, _ = fmt.Fprintf(e.output, " %s", name)
-		}
-		_, _ = fmt.Fprintf(e.output, "\n")
-	}
-
-	// Build and execute the actual command
-	return e.buildGitCommand(operation, resource, name, options, false)
 }
 
 // buildGitCommand builds and displays the git command
@@ -1284,66 +1079,6 @@ func (e *Engine) buildGitCommand(operation, resource, name string, options map[s
 	// return cmd.Run()
 
 	return nil
-}
-
-// executeHTTP executes HTTP operations
-func (e *Engine) executeHTTP(httpStmt *ast.HTTPStatement, ctx *ExecutionContext) error {
-	// Interpolate variables in HTTP statement
-	method := httpStmt.Method
-	url := e.interpolateVariables(httpStmt.URL, ctx)
-	body := e.interpolateVariables(httpStmt.Body, ctx)
-
-	// Interpolate headers
-	headers := make(map[string]string, len(httpStmt.Headers))
-	for key, value := range httpStmt.Headers {
-		headers[key] = e.interpolateVariables(value, ctx)
-	}
-
-	// Interpolate auth
-	auth := make(map[string]string, len(httpStmt.Auth))
-	for key, value := range httpStmt.Auth {
-		auth[key] = e.interpolateVariables(value, ctx)
-	}
-
-	// Interpolate options
-	options := make(map[string]string, len(httpStmt.Options))
-	for key, value := range httpStmt.Options {
-		options[key] = e.interpolateVariables(value, ctx)
-	}
-
-	if e.dryRun {
-		return e.buildHTTPCommand(method, url, body, headers, auth, options, true)
-	}
-
-	// Show what we're about to do with appropriate emoji
-	switch method {
-	case "GET":
-		_, _ = fmt.Fprintf(e.output, "📥 GET request to: %s\n", url)
-	case "POST":
-		_, _ = fmt.Fprintf(e.output, "📤 POST request to: %s\n", url)
-	case "PUT":
-		_, _ = fmt.Fprintf(e.output, "🔄 PUT request to: %s\n", url)
-	case "PATCH":
-		_, _ = fmt.Fprintf(e.output, "🔧 PATCH request to: %s\n", url)
-	case "DELETE":
-		_, _ = fmt.Fprintf(e.output, "🗑️  DELETE request to: %s\n", url)
-	case "HEAD":
-		_, _ = fmt.Fprintf(e.output, "🔍 HEAD request to: %s\n", url)
-	default:
-		_, _ = fmt.Fprintf(e.output, "🌐 %s request to: %s\n", method, url)
-	}
-
-	// Handle special HTTP operations
-	if downloadPath, exists := options["download"]; exists {
-		_, _ = fmt.Fprintf(e.output, "💾 Downloading to: %s\n", downloadPath)
-	}
-
-	if uploadPath, exists := options["upload"]; exists {
-		_, _ = fmt.Fprintf(e.output, "📤 Uploading from: %s\n", uploadPath)
-	}
-
-	// Build and execute the actual HTTP request
-	return e.buildHTTPCommand(method, url, body, headers, auth, options, false)
 }
 
 // buildHTTPCommand builds and displays the HTTP request details
@@ -1529,26 +1264,6 @@ exit 1`, timeout, retryInterval, target)
 	// return cmd.Run()
 
 	return nil
-}
-
-// executeDetection executes smart detection operations
-func (e *Engine) executeDetection(detectionStmt *ast.DetectionStatement, ctx *ExecutionContext) error {
-	detector := detection.NewDetector()
-
-	switch detectionStmt.Type {
-	case "detect":
-		return e.executeDetectOperation(detector, detectionStmt, ctx)
-	case "detect_available":
-		return e.executeDetectAvailable(detector, detectionStmt, ctx)
-	case "if_available":
-		return e.executeIfAvailable(detector, detectionStmt, ctx)
-	case "if_version":
-		return e.executeIfVersion(detector, detectionStmt, ctx)
-	case "when_environment":
-		return e.executeWhenEnvironment(detector, detectionStmt, ctx)
-	default:
-		return fmt.Errorf("unknown detection type: %s", detectionStmt.Type)
-	}
 }
 
 // executeDetectOperation executes detect operations
